Modfied structure to accomodate libraries to do different tasks and moved initating a device to be part of new library
