+++ blockify/Cargo.toml
@@ -1,6 +1,5 @@
-workspace = { members = ["blockify"] }
 [package]
-name = "rust-craft"
+name = "blockify"
 version = "0.1.0"
 edition = "2021"
 
@@ -8,5 +7,3 @@ edition = "2021"
 
 [dependencies]
 vulkano = "0.34.0"
-blockify = { path = "blockify" }
-
