+++ blockify/src/lib.rs
@@ -1,0 +1,48 @@
+use vulkano::VulkanLibrary;
+use vulkano::instance::{Instance, InstanceCreateInfo};
+use vulkano::device::QueueFlags;
+use vulkano::device::{Device, DeviceCreateInfo, QueueCreateInfo};
+use vulkano::memory::allocator::StandardMemoryAllocator;
+
+pub fn get_devices(left: usize, right: usize) -> usize {
+    
+    let library = VulkanLibrary::new().expect("no local Vulkan library/DLL");
+    let instance = Instance::new(library, InstanceCreateInfo::default())
+        .expect("failed to create instance");
+    let physical_device = instance
+        .enumerate_physical_devices()
+        .expect("could not enumerate devices")
+        .next()
+        .expect("no devices available");
+
+
+
+    let queue_family_index = physical_device
+        .queue_family_properties()
+        .iter()
+        .enumerate()
+        .position(|(_queue_family_index, queue_family_properties)| {
+            queue_family_properties.queue_flags.contains(QueueFlags::GRAPHICS)
+        })
+        .expect("couldn't find a graphical queue family") as u32;
+
+        
+
+    let (device, mut queues) = Device::new(
+        physical_device,
+        DeviceCreateInfo {
+            // here we pass the desired queue family to use by index
+            queue_create_infos: vec![QueueCreateInfo {
+                queue_family_index,
+                ..Default::default()
+            }],
+            ..Default::default()
+        },
+    )
+    .expect("failed to create device");
+
+    let queue = queues.next().unwrap();
+    return left + right;
+}
+
+
